back to predecalre ang0,...

+++ examples/TTI_codegen.py
@@ -14,7 +14,7 @@ class TTI_cg:
         self.t_order = t_order
         self.s_order = s_order
         self.data = data
-        self.dtype = np.float64
+        self.dtype = np.float32
         self.dt = model.get_critical_dt()
         self.h = model.get_spacing()
         self.nbpml = nbpml
